#5 Implement visualization
This commit completes the 35-core graph visualization.

There are drop down boxes to select the start and end nodes of the current path.

The current path is highlighted.

The start and end nodes are shown in different colors.

There's a right hand panel to show the current path.

We can drag nodes around.

There's node hovering effects.

There are many other features as well.